ログイン・管理者ログインのコンポーネント分け、ボタン活性化実装
