+++ data/db.json
@@ -5,7 +5,7 @@
       "firstName": "",
       "lastName": "",
       "email": "test@test.com",
-      "password": "test1234",
+      "password": "Test1234",
       "isAdmin": false,
       "authId": "",
       "polledPopular": false,
@@ -40,7 +40,7 @@
       "name": "ブライトブレンド",
       "description": "ミディアムローストの豆をブレンドしたブライトブレンドは、キャラメル、ベリー、はちみつのバランスのとれたほんのり甘い香りが楽しめる一杯です。",
       "image": "/item.png",
-      "itemCategory":2,
+      "itemCategory": 2,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -51,7 +51,7 @@
       "name": "LAVAZZA CLASSICO",
       "description": "しっかりとした珈琲感とドライフルーツの風味が特徴のミディアムローストコーヒー。バランスのとれたリッチな味わいがお好みの方へオススメです。",
       "image": "/item.png",
-      "itemCategory":2,
+      "itemCategory": 2,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -62,7 +62,7 @@
       "name": "LAVAZZA INTENSO",
       "description": "スモーキーで香ばしい力強いフレーバーのダークローストコーヒー。キャラメリゼしたような風味と濃厚な味わいはリアルミルクフロスとも相性抜群です。",
       "image": "/item.png",
-      "itemCategory":1,
+      "itemCategory": 1,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -73,7 +73,7 @@
       "name": "プレミアム スムースロースト",
       "description": "ほど良いコクと軽やかで優しいアフターテイスト。一日のどんな気分にも寄り添う、プレミアムにスムーズな一杯です。",
       "image": "/item.png",
-      "itemCategory":3,
+      "itemCategory": 3,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -84,7 +84,7 @@
       "name": "アメリカンロースト",
       "description": "口当たりの良い さわやかな酸味。ほのかな甘いノートが最後に残ります。",
       "image": "/item.png",
-      "itemCategory":3,
+      "itemCategory": 3,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -95,7 +95,7 @@
       "name": "コロンビア",
       "description": "100%南米コロンビア産の生豆使用。香ばしい香りと酸味・甘味・コクのバランスはマイルドそのものです。",
       "image": "/item.png",
-      "itemCategory":2,
+      "itemCategory": 2,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -106,7 +106,7 @@
       "name": "スマトラ",
       "description": "100%スマトラ島産アラビカ豆。火山性土で育った豆からの力強い個性的なアロマは、複雑で深いコク、そしてほのかな甘みへと続きます。",
       "image": "/item.png",
-      "itemCategory":1,
+      "itemCategory": 1,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -117,7 +117,7 @@
       "name": "プレミアスムースデカフ",
       "description": "市場のニーズが高まっているデカフェコーヒー。フルーティーな軽さの中に程よいコクがアフターテイストを満たします。",
       "image": "/item.png",
-      "itemCategory":4,
+      "itemCategory": 4,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -128,7 +128,7 @@
       "name": "イングリッシュ ブレックファースト",
       "description": "バランスのよい渋味。しっかりとした味わい豊かな紅茶。",
       "image": "/item.png",
-      "itemCategory":5,
+      "itemCategory": 5,
       "createdAt": "",
       "intheOffice": false,
       "author": "",
@@ -139,7 +139,7 @@
       "name": "アールグレイ",
       "description": "爽やかなベルガモットの香りとほどよい風味が特徴的なお茶。シトラスとフルーツの香りとほのかに感じるスモーキーな香りのブレンドを楽しんでいただけます。",
       "image": "/item.png",
-      "itemCategory":5,
+      "itemCategory": 5,
       "createdAt": "",
       "intheOffice": false,
       "author": "",
@@ -150,7 +150,7 @@
       "name": "レモンハーブ",
       "description": "軽やかなレモンと柑橘系の香りに、フローラルなハチミツの風味が華を添えるノンカフェインのハーブティーです。",
       "image": "/item.png",
-      "itemCategory":5,
+      "itemCategory": 5,
       "createdAt": "",
       "intheOffice": false,
       "author": "",
@@ -161,7 +161,7 @@
       "name": "ジャスミン茶",
       "description": "ジャスミンの芳香を採りいれた渋みの少ないさっぱりとした緑茶です。",
       "image": "/item.png",
-      "itemCategory":5,
+      "itemCategory": 5,
       "createdAt": "",
       "intheOffice": false,
       "author": "",
@@ -172,7 +172,7 @@
       "name": "緑茶",
       "description": "豊かな香りとコク、まろやかな風味。深蒸し煎茶から引き出される ほど良い渋味と旨味は日本人の好みにぴったり。 国産茶葉100%使用。",
       "image": "/item.png",
-      "itemCategory":5,
+      "itemCategory": 5,
       "createdAt": "",
       "intheOffice": false,
       "author": "",
@@ -183,7 +183,7 @@
       "name": "ホットココア",
       "description": "口当たりがよく、滑らかな舌触りのココアです。コク深いミルクチョコレートの風味は、ほっと一息つくリラックスタイムにピッタリです。女性にうれしい低カロリードリンク。（1杯あたり66.1kcal）",
       "image": "/item.png",
-      "itemCategory":6,
+      "itemCategory": 6,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -194,7 +194,7 @@
       "name": "リアルミルクフロス",
       "description": "原材料牛乳100%。砂糖や香料不使用のラテ用ミルクパウダーです。ミルク風味豊かでふわふわのカフェラテ・ティーラテ・チャイラテ・ココアラテがお楽しみいただけます。",
       "image": "/item.png",
-      "itemCategory":7,
+      "itemCategory": 7,
       "createdAt": "",
       "intheOffice": true,
       "author": "",
@@ -203,77 +203,80 @@
   ],
   "itemCategory": [
     {
-      "id":1,
-      "category":"コーヒー",
-      "name":"ダーク（深煎り）"
+      "id": 1,
+      "category": "コーヒー",
+      "name": "ダーク（深煎り）"
     },
     {
-      "id":2,
-      "category":"コーヒー",
-      "name":"ミディアム（中煎り）"
+      "id": 2,
+      "category": "コーヒー",
+      "name": "ミディアム（中煎り）"
     },
     {
-      "id":3,
-      "category":"コーヒー",
-      "name":"ライト（浅煎り）"
+      "id": 3,
+      "category": "コーヒー",
+      "name": "ライト（浅煎り）"
     },
     {
-      "id":4,
-      "category":"コーヒー",
-      "name":"カフェインレス"
+      "id": 4,
+      "category": "コーヒー",
+      "name": "カフェインレス"
     },
     {
-      "id":5,
-      "category":"ティー",
-      "name":""
+      "id": 5,
+      "category": "ティー",
+      "name": ""
     },
     {
-      "id":6,
-      "category":"ココア",
-      "name":""
+      "id": 6,
+      "category": "ココア",
+      "name": ""
     },
     {
-      "id":7,
-      "category":"その他",
-      "name":""
+      "id": 7,
+      "category": "その他",
+      "name": ""
     }
   ],
   "questionnaire": [
     {
       "id": 1,
       "name": "社内drink人気投票",
-      "polledItems":[{"id":1,"itemId":1,"detail":"","result":""},{"id":2,"itemId":2,"detail":"","result":""}],
-      "createdAt":"2023-04-1",
-      "category":1,
-      "startDate":"2023-05-1",
-      "endDate":"2023-05-1",
-      "author":1
+      "polledItems": [
+        { "id": 1, "itemId": 1, "detail": "", "result": "" },
+        { "id": 2, "itemId": 2, "detail": "", "result": "" }
+      ],
+      "createdAt": "2023-04-1",
+      "category": 1,
+      "startDate": "2023-05-1",
+      "endDate": "2023-05-1",
+      "author": 1
     },
     {
       "id": 2,
       "name": "メニューの広がり投票",
-      "polledItems":[{"id":1,"itemId":1,"detail":"","result":""}],
-      "createdAt":"2023-04-1T02:50:27.944Z",
-      "category":2,
-      "startDate":"2023-05-1T02:50:27.944Z",
-      "endDate":"2023-05-1T02:50:27.944Z",
-      "author":1
+      "polledItems": [{ "id": 1, "itemId": 1, "detail": "", "result": "" }],
+      "createdAt": "2023-04-1T02:50:27.944Z",
+      "category": 2,
+      "startDate": "2023-05-1T02:50:27.944Z",
+      "endDate": "2023-05-1T02:50:27.944Z",
+      "author": 1
     }
   ],
   "polls": [
     {
       "id": 1,
       "questionnaireId": 1,
-      "userId":2,
-      "result":1,
-      "createdAt":"2023-05-2T02:50:27.944Z"
+      "userId": 2,
+      "result": 1,
+      "createdAt": "2023-05-2T02:50:27.944Z"
     },
     {
       "id": 1,
       "questionnaireId": 1,
-      "userId":1,
-      "result":2,
-      "createdAt":"2023-05-3T02:50:27.944Z"
+      "userId": 1,
+      "result": 2,
+      "createdAt": "2023-05-3T02:50:27.944Z"
     }
   ],
   "contact": [],
